Bardiches have a 1 in 100 chance of beheading monsters.
This idea was pitched by disperse.

The player must be (a)pplying the bardiche to pound it for this effect to work. This has the same restrictions as Vorpal Blade, so it cannot behead headless, noncorporeal, or amorphous monsters. You also cannot behead while engulfed.

Monsters also have this advantage on the player...