Spetums can skewer up to 3 monsters.
Suggested by disperse.

This ability is quite powerful, so it can trigger only when the player is skilled or better in polearms and when they are riding a steed. It will not trigger when merely pounding - that would also make the implementation a lot more complicated.

I updated the previous should_skewer method to accomodate more monsters in a row. So if there are any peaceful monsters in the 2 squares behind the main target, the extra skewer will not trigger unless the weapon is cursed.

Another update to the spear skewering is that the skewer fails to continue if any of the hits miss. So, every monster must be hit/skewered for the loop to continue.

Because monsters cannot ride steeds - we also don't have to worry about implementing this on the monster side!

Roles that can currently take advantage of this mechanic are: rangers, samurai, valkyries, and knights. In a followup commit we will grant some riding skill to priests so they can also use it.