Fix: heap-use-after-free after nymph steals chain. Fixes #10.
I thought this had to do with a nymph removing an amulet of flying or boots of flying when we were over water, but turns out it was a null pointer being referenced. I was able to reproduce the problem by having a nymph steal my chain (while punished) when my inventory was completely empty.

After being unpunished, the code in the function remove_worn_item would run down to this:
obj->in_use = oldinuse;
and hit the heap-use-after-free.

For some reason I could not reproduce this in Vanilla 3.7 even with ASAN enabled. But the fix seems to work here.