Amulets of magical breathing are immune to water damage.
They would be fairly useless if they just rusted away in the water...

I only added protection for water rusting, not any other strange types (like rust monster passive attacks), so there could be other avenues of damage. They can also still be corroded.

Note that these will not display 'rustproof' in the inventory so that their identity is not revealed. You can employ a trick to identify them by dipping them in a water source. If it is any other amulet it will rust and these of course, will not.