Samurai get a bonus for twoweaponing a katana with a wakizashi.
This is based on a YANI by aosdict (https://nethack-yanis.github.io/yanis/3390.html).

They get a flat +1 to-hit and a little damage bonus based on their skill in two-weaponing. So if you have basic in two-weaponing you get +d2 damage, skilled would be +d3, etc.

One of my favorite books is the Book of Five Rings by Miyamoto Musashi. Since this technique is based on that I added a message for when samurai begin twoweaponing these two weapons "You begin the Two Heavens technique." :)