Ported the hint system from UnNetHack.
With this system a helpful engraving appears on the first level in the starting room (with the upstair). I added a bunch of hints regarding mechanics changes. I also made it possible for these to appear in levels 2-4 with lessening frequency.

I didn't port the hint daemon - it's fun but we'll keep it simple.