Handle morning star and flail stunning in hitmu. Fixes #28.
When polymorphed into a form with a passive erosion, we could passively erode a monster's weapon away and it could be referenced after being destroyed.

When monster's were hitting us with morning stars or flails, the handling could result in that check after their weapon got destroyed. I was able to relocate that code to a more suitable place before the erosion takes place.