Barbarians get a blood rage bonus for low health.
This is adapted from SpliceHack's Blood Rage skill. Here, we don't use any skill system or points, the barbarian just gets it baked in for free.

When a barbarian is below 50% health, they get a damage bonus. When below 25% health, it's doubled. If all the conditions are met, the bonus triggers 2 out of 3 times.

This only works for melee attacks and when the barbarian is in their base form (not polymorphed). It also only becomes available after level 3 so the first few levels can't be steamrolled.

The bonus is simply: (your level divided by 5) + 1.

So if you are level 5, you have 40% health. It's 1. If you have 20%, it's 2.

I added some more messages to describe the attack when it triggers. I think this will be nice to add a bit of symmetry to the monster's berserk behavior. It encourages barbarians to gain as many levels as possible and maybe to intentionally lower their HP at strategic times. It also gives barbs a fighting chance when their backs are against the wall - I hope it plays out well.