Crossbows no longer grant multishot, instead increasing damage.
From dNetHack. I like this change because it makes crossbows much beefier, especially for gnomes. In dnh, gnomes get an extra die of damage, I'm not sure how much but I went with d3 just so that they aren't too strong early game. Plus, if we apply this for monsters we don't want them getting too big a boost either.

This also means you don't need to fire as many crossbow bolts to do roughly the same amount of damage. This means less mulching, and less collecting. It also means that it's more difficult to clear rows of enemies - unlike a flurry of multishot arrows, if the first monster is killed that is it, nothing behind will be hit.

Needs a followup commit to grant all gnomes crossbow skill.