Revert "Water Elementals can now engulf and suffocate."
This reverts commit 4bd835c7fbb8927ed923991d0a6f1c253a0c3857.

Allowing this attack for water elementals is simply too powerful. There is little to no defense against it, and the plane of water is overrun with water elementals. The player can wear an amulet of magical breathing to survive the suffocation attack, but all of the metallic items are subject to rusting and destruction. Also - this attack has proven buggy and a couple of things still cause strangulation to persist even after exiting the engulf attack. I think removing it is for the best, there is a reason xNetHack followed the same path.

I am still keeping the passive rusting attack, that is thematic and a minor buff.