Fix: Blocking attacks could reveal the item type.
In commit a1b275418, I attempted to make the blocking messages a bit less wordy, but ended up revealing unknown items in the process. So if you blocked an attack with your unknown fencing gloves, it would reveal them as ie: gauntlets of power. This commit simply reverts the original commit.