Slipping in grease traps can cause you or monsters to become paralyzed for a few turns and inflicts a small amount of damage.
This is something I meant to add earlier. The handling is still a bit choppy in places. Steeds are still exempt from damage and stunning - maybe they only slip and buck you off? So some of those details remain. But this definitely makes the trap a bit more dangerous to run into or to try and abuse to gain a bunch of greased items. On the other hand, monsters can be lured into it now making it more useful...