Implement spell beings and summon cards in conjunction with the cartomancers abilities.
This is a fairly complex commit, so there will probably be some bugs/mistakes surfacing later.

Spell beings: These originated with SLASH'EM with the flame spheres and freezing spheres, and appear here in a similar form. The lifespan of these is shorter though.
* Spell beings are killed if they are canceled.
* Spell beings can not leave corpses.
* Spell beings can not leave summon cards.
* If monsters leave a summon card (when defeated by a cartomancer), they will not leave a corpse.

A small detail to tend to later is how much XP is given to the player when a spell being is killed. It should be 1, but the transfer between Hack'EM and 3.7 wasn't clear.

The summon cards drop when cartomancers kill monsters.

In this commit I also adjusted the Carto's quest artifact to spew out a bunch of spell beings when invoked.

I also added the hackem fix for merging summon cards: 03b0cec7d Fix: Summon cards were merging incorrectly with other cards. This was bug was also present in SpliceHack (https://github.com/NullCGT/SpliceHack/issues/232).