Cartomancers can throw summon cards to activate them.
Ported from Hack'EM: The cartomancer can throw their summon cards to activate them. This was adapted from the spicycat's idea on the NetHack Ideas Archive (https://nethack-yanis.github.io/yanis/3867.html). If the card hits a monster, it will activate on a square next to the monster, otherwise it will activate wherever it lands.

I also consolidated the additional changes of throwing up or down that came later in hackem.