New card types for cartomancer: zappable cards. These can also drop when cartomancers kill monsters. They act as one-use wands. These have a 0% generation chance so that other roles will not see them randomly, but they could see them in bones.
All wand types are eligible for these zaps except wishing and nothing. It's possible we might want extremely rare drops to be wishing - not sure about this yet...

Throwing or zapping these cards will also activate them. However, unlike wands, they are not subject to the cursed backfire chance and they can't be (a)pplied to break them.

Hopefully these add some fun variety to the cartomancer death drops and give them some extra resources to work with.

Ported from Hack'EM as a new feature for the cartomancer.