Cartomancers can receive razor card death drops.
This should make the game a bit more stream-lined, and may even setup the role to be
stripped of any starting weapon at all!

Currently there is a 1 in 20 chance of a stack of razor cards (booster pack??) when a
drop occurs, but this will probably be adjusted quite a bit in the future depending on
how it feels. I adjusted the drop-rate of zap cards to be 4 in 20 (when a drop occurs).