Vampires get infravision.
I have two motivations for this change. The first is to make vampires stand out a bit more from the other races. One of the tougher aspects of NerfHack is that no starting races get infravision anymore, but it feels right for vampires to get it. The second is that classically from movies I think they always have the 'predator vision' and this ability fits well with them.