Ported the Temple to Moloch level from SLASH'EM.
Appears from level 20-30. Also added two ghoul mages in the back as a difficulty check.

I think this and the Lost Tomb will act as good additional level content. We won't have much more but these levels add a lot of items and loot for the player to go through. This level also offers a guaranteed uncracked altar that the player can attempt to convert.