The ring of levitation is a valid starting ring.
There is a comment that says overly powerful items are not permitted for starting inventory. The =oLev doesn't seem to be that powerful anymore. I'm guessing that it's power derived from the ability to levitate over dust Elbereth and fight without fear. This mechanic has been nerfed so we'll open it up again.