Levels of erosion on an object can affect its price (from EvilHack).
Commit 47d3518f393:

'Depending on the level of erosion on an object will effect its price,
both buying from and selling to a shopkeeper. If the object is 'fixed',
that will affect the price also. The breakdown:

* one level of erosion, price is decreased by 1.5x the base value
* two levels of erosion, price is decreased by 2x the base value
* three levels of erosion, price is decreased by 3x the base value
* 'fixed' objects, price is increased by 2x the base value

There are two types of erosion - oeroded and oeroded2. These can stack,
so if you have an object that is both rusty and corroded, it's value
will decrease significantly.

'Fixed' refers to objects that are purposely made erodeproof. Objects
made of materials that are already inherently erodeproof are not
considered 'fixed'.'

--hackemslashem: Because we have a lot more eroded and erodeproof items in play, this commit works really well, kudos to K2 for coming up with this!