Skewering doesn't trigger most passive attacks.
Another thing that popped up with this semi-ranged attack is that some monsters would trigger passive attacks even if we were not next to them. Since we are separated by a square (and usually another monster), this doesn't make sense. It only makes sense for the passive electricity when we attack with a metal weapon.