Add rivers to some quest fill maps.
This is a nice alternative to rivers in the mines, which some players thought were placed a bit early for the game. Here, if a quest level has standard mines fill using solid rock, rivers will be placed. I added a mines style level to the monk quest so they could enjoy it as well.

Not all roles get this since some of their levels are ice/water already. But the ranger is an interesting case since their mines fill style uses trees, so they will get rivers in between the tree fill.

In the future, it would be nice to port this to LUA instead of jamming it in mkmap.c.