Breaking rocks is now an occupation.
After playing a lot of caveman with the new rock breaking mechanics from xnh and evilhack, I have grown to loathe applying rocks to each-other. In this commit, we automatically start a rock breaking occupation when rocks are applied. Occupations naturally stop when threats approach, making this much safer than the mindless button mashing before!

I simply copied the way SLASH'EM implemented the whetstone occupation. There is a separate rock_breaking object which feels a little janky, but for simplicity I think it works for now and is easy to understand in the code. There is no prompt for the occupation, maybe we'll need one if it proves too dangerous to just jump into the occupation but for now lets leave it.