Allow steadfastness to work properly.
I noticed in my playtesting that sometimes while wearing Pridwen (which now grants steadfastness), that I was still getting knocked around. My first discovery was that there was an exception for flying or levitation, then you were not being protected because supposedly the force would be enough to knock you through the air. However, I also discovered that I was being knocked on the plane of water! Looking closer, there was even a mini-exception because you are floating in a bubble. These exceptions seem a bit extreme and unintuitive. They also destroy the effectiveness of anything that should grant steadfastness.

We will employ a little hand-waving so that items that grant steadfastness do so under ALL circumstances. Because this list is so limited (Loadstones, Giantslayer, Pridwin, and Load Brand) I don't think it's 'game breaking' and it lets these items live up to their full potential.