Weapons can be enchanted much higher, with a soft limit of +11.
This is lightly adapted from Gnollhack, with my own implementation. I've liked the idea of giving the player more power in the late-game, and this seems like the most free way to do it. Basically, the rules work the same as before, but instead of the soft-limit being 5, it's now 11. So one can read a scroll of enchant weapon at +11 safely, hoping to get a +13 weapon, but probably getting a +12 one most of the time.

I had to tweak the amount that is granted slightly, which required hard-coding a special function. Enchantment is rewarded in this fashion:
* +0 - +3 weapon: Can receive d3 enchantment.
* +4 - +11 weapon: Can receive d2 enchantment.
* Anything over +11 can only receive +1.

I think with the nerfs that have been put in place, this makes something like this maybe actually kind-of balanced. Players already have a hard time finding markers and wishes, and polypiling is much less effective for hoarding enchant scrolls. Players will still have to consider carefully what they invest their scrolls in - I'm excited to see it at work.