Wand explosions discharge their effects (SLASH'EM).
This is another neat - but deadly - quirk from SLASH'EM. It's actually a little surprising that vanilla hasn't adopted this. Basically, wand effects only occurred when the hero purposely broke (applied) the wand. When a wand exploded in other fashions, these didn't occur (zapping, backfire, etc).

My initial implementation is not as concise as I'd like, there is some code duplication, but I tried to consolidate where I could.

One wrinkle is that the damage was different for manually breaking a wand vs having it explode. I erred on the side of sticking with the formula Vanilla used - it's pretty nasty damage wise, and more dangerous than SLASH'EM's formula.