Mon zapping cursed wands also explode.
This should make things a little more entertaining. This follows in the footsteps of the SLASH'EM mechanic where exploding wands blast their effect. In SLASHEM, muse didn't do that meaning the player was the only one penalized. Here, we have increased the chance so that monsters have double the chance of exploding a wand. This could be good or bad for the player...