Extract wizard's tower to it's own branch.
With a lot of help from xNetHack's implementation of this, I have wanted the wizard's tower on it's own branch so it makes the dungeon a little shorter. After experimenting with the tower levels just after Medusa's, it felt a bit awkward but it was a step in the right direction.

I also spent some time labeling the dungeons and levels in decl.c because otherwise it's an unreadable mess.