Port Amulet of Storms (from xNetHack).
I might jumping-the-gun a little here since this comes from an unreleased version of xnh, but this item looks really neat!

From commit 37b10da9b58:

'Note by copperwater: Intent of this is to continue expanding a few pools
of artifacts:
    - non-weapon non-quest artifacts
    - chaotic artifacts
    - less powerful artifacts

while still being useful enough to appreciate when given as a sacrifice
gift.

From the original patch, I adjusted the price down (it's not powerful
enough to be worth 2000 base price) and refactored is_stormy_monster for
clarity. I also gave it an additional side effect of blocking lightning
paralysis (but not damage) on the Plane of Air; seems reasonable that
it'd have this protective effect as well.'