Handle tame monsters going berserk. Fixes #32.
I was seeing some weird stuff come up in the fuzzer with hostile tame monsters. I thought it was the cartomancer, but it was berserking monsters.

When a tame monster goes berserk, it needs some additional checks - so I used some of the checks that SLASH'EM's traitorous monsters get. I also added a much lower HP threshold of below 1/6th the pet's HP. Although pets should be able to berserk, this untames them, and should not trigger nearly as often as it does for regular berserkers.