Scrolls of knowledge (from SpliceHack).
My main motivation for porting this scroll was another $50 scroll to compete with scrolls of light.

This scroll is a bit different from the Splice implementation. I've made it a bit more streamlined. Instead of the user typing in the name to formally identify - they learn about a random magic item or two. This reduces the abusability of the item a bit, but probably makes it a little more fun. The exception is archeologists - who *do* still get to specify exactly what they want to learn about in addition to getting the random knowledge.

Because of the soft nerf, I've allowed this scroll to be written.