Port scroll of water from xNetHack.
Which was ported from UnNetHack, known there as the scroll of flood. I do like scroll of water better, provides some symmetry to the scroll of fire.

xnh removed the removed the possibility of putting water on your own square - which was a core mechanic of the scroll in UnNetHack. I re-added this, but otherwise the scroll remains the same as in xnh.

I mostly want this to add another $200 scroll. It's not super game changing, but adds a little variety.