Scroll of cloning (from SpliceHack).
This will probably be the last new scroll I import for NerfHack. I had some reservations about bringing it in but it is interesting, and I don't think it's too abusable with the built in restrictions.

Cloning these objects has nerfed results (obviously for abuse or balance considerations):
* magic marker -> athame
* magic lamp -> oil lamp
* wand of wishing -> wand of wonder
* bell of opening -> bell
* candelabrum of invocation -> wax candle
* amulet of yendor -> cheap plastic imitation amulet
* book of the dead -> blank spellbook
* scroll of cloning -> blank scroll (prevents infinite cloning loops)
* artifacts all copy the base item, but lose the name and the status of being an artifact.

To copy the enchantment on an item, the scroll must be blessed.
If the scroll is cursed, the resulting item will be cursed.

All other properties on the item should be copied exactly as is. This enables fun tricks like cloning that +13 fixed crysknife, or that blessed figurine of an Archon you've been holding onto.

When confused, the scroll clones the player. The resulting clone will have no inventory. A blessed scroll creates a tame clone, an uncursed scroll makes a peaceful clone, and a cursed scroll makes a hostile clone. An interesting side-effect of this is that cloning yourself reduces your current HP by half.

Couple updates:
* Unpaid items cloned in shops become property of the shopkeeper.
* More item properties are carried over: erosionproofing, container status, and some other misc things to make sure items are truly cloned.
* Intelligent monsters can use these scrolls to clone themselves!  I think this sort of balances out the scroll a bit, since a good number will be used up by monsters. When monsters read them, they always act as confused - so they can only clone themselves. It's a bit weird, but it's much easier than allowing them to clone items and needing to create a system for that mess...
* Unique monsters can also use these scrolls... The Wizard of Yendor still has to go through the Double Trouble routine, so no more than 2 Rodneys will ever oppose you. But it's possible for other uniques to clone themselves more than once if the opportunity arises.