Soft revert 82f0b1e8e - Scared hostile monster which cannot move away will attack.
Reason: Having monsters attack while scared or fleeing literally breaks the mechanic for the scroll of scare monster and how Elbereth works. I strongly disagree with this change in 3.7 and I think a lot of players are unpleasantly surprised by it when they discover it. Elbereth and Scare Monster have already been nerfed to death, I'm not sure we need to break those mechanics any further.

This just disables the "fleeing" attack for scared monsters. Scrolls of scare monster and Elbereth should protect the player from melee range attacks, but not ranged.