The caveman's gods can also ignore major trouble.
When porting over the new behavior for caveman gods when #praying, I noticed that the major troubles didn't have the same treatment as the minor trouble. To make things more consistent (ie: "mean"), we also are giving a 1 in 13 chance for major troubles to be ignored. This means cave people cannot rely on their gods for all their help, there must be backup plans!