Wishes for artifacts only count if the player actually got the artifact.
In response to some conversation in #evilhack about wishes and the probabilities going down for artiwishes that were unsuccessful. In this variant, wishes and artifacts are very scarce. I think we can lighten this restriction so it doesn't punish so heavily for artiwishes.

I think I got the implementation correct, in conduct it will only show your successful artiwishes.