Jewelery eating only grants temporary trinsics.
This is a further resource constriction nerf to push each role/race to utilize it's individual strengths rather than being able to attain all the trinsics easily by eating them. It's also intended to make the actual items more valuable and not as disposable. I was also not satisfied with the state of temp-intrinsics as it was started with see invisible/invisibility/etc, I think it should be consistent across the board or not at all.

To compensate for the temporary nature of these intrinsics, the possibility of actually getting the property has been inversed.
* Before, the chance of getting an intrinsic from a ring by eating it was 1 in 3. It has been changed to 2 in 3.
* Before, the chance of getting an intrinsic from an amulet by eating it was 1 in 5. It has been changed to 4 in 5.

Summary of revised eating effects:
- Eating a ring now grants 750-1500 turns of it's property intrinsically.
- Eating an amulet now grants 1250-2000 turns of it's property intrinsically.
- Eating the amulet of reflection now gives intrinsic reflection for 1250-2000 turns.
- Eating the amulet of flying now gives intrinsic flying for 1250-2000 turns.

Some accessories still don't do anything at all: Amulet of life saving, ring of sustain ability.