Cartomancer card abuse penalty revamp.
Instead of summoning kops - we are going to enforce some pretty heft penalties to keep the card gods respect:
- writing cards: -20 alignment, -5 luck
- cloning cards: -20 alignment, -5 luck
- poly zap/dip of cards: -10 alignment, -2 luck

If we just summon kops, they can be dealt with pretty easily and it's more of a joke than an actual solution to making this a serious part of the Cartomancer weak. We basically want to get cartomancers to use the cards that they are given or find through the card drop mechanics. I've left out the 'abuse' of cancelling or destroying cards for now. It's more involved to assign responsibility for who or what destroyed a card and there are potentially many ways to destroy/dilute cards. The important part is forgery, a cardinal sin for card duelists!

If a player decides they want to go ahead and forge cards anyway - they will have to deal with some very nasty Erinyes.