Altar artifact gift frequency depends on level (from EvilHack).
I think this is a good complement to the difficulty scaling. I've adapted the gifting schedule from EvilHack, but not the non-artifact gifts.

As your level increases, the probability of receiving an artifact increases. Under level 4 is not possible:
    * level  4: 10% chance
    * level  9: 20% chance
    * level 12: 30% chance
    * level 14: 40% chance
    * level 17: 50% chance
    * level 19: 60% chance
    * level 21: 70% chance
    * level 23: 80% chance
    * level 24: 90% chance
    * level 26 or greater: 100% chance

This rewards leveling up and slows down the power grab that some characters might go for with a guaranteed sacrifice gift (ie: Barbarians). It also slows down the rate at which altars are cracking and getting destroyed, giving more time for players to utilize the altars for other things like building Luck or getting crowned.

A key difference from EvilHack is that the number of artiwishes is not factored in when determining the chance of a gift. The previous chances we had worked out still apply, just filtered through the level check.