Forging recipes (from EvilHack).
I took a while to think about how forging would work in NerfHack. For a time, I didn't really think it had a place, but I realized a more restrained approach could work. So anyway, we do have forging recipes with the following guidelines:
* No artifact forging
* No forging armor
* Most of the forgeable weapons are blades and blunt. I thought being able to forge anything intricate (anything with chains) or precision based (like a scalpal) felt a bit too out of the realm of possibility so those were left out.
* Only samurai can forge katanas and tsurugi. No other restrictions on other items.
* Instead of cavemen dipping flint into a forge for sling bullets, we now have the evilhack forge recipe.
* Forging recipes are displayed in the object lookup.

Aside from the recipe changes - the biggest change is that we no longer have a #forge or #craft command. Simply get a hammer and (a)pply it when standing over the forge. I think this is a much easier way to use the forge and reduces the work of creating a #forge command.