Cancelled unicorn horns become degraded, unusable for curing.
Adapted from YANI #743 by jonadab (https://nethack-yanis.github.io/yanis/743.html).

This one is a little nasty, but makes perfect sense. Most degraded unicorn horns come from revived unicorns that were slaughtered. But if a unicorn horn is cancelled, it's also degraded. If a player attempts to cure themselves with a degraded unihorn, it is useless for curing ailments, and 1 out of 10 times it will simply crumble.

For this reason, we will probably not ever implement the monster spellcasting cancellation spell. That would be too cruel. But there are still monsters with wands of cancellation, and if one explodes on us...