The Sokoban prize can also be a cloak of magic resistance or a magic marker.
This kind of follows in the footsteps of other variants, but more simplified and much less moving parts in regard to the multiple-doors approach. It's not uncommon to get one or both the amulet of reflection and the bag of holding, so this improves the chances you will get a new and desireable item for completing Soko.