Cure sickness is now directional (from EvilHack).
From commit 8afcbbb7f8a. I adapted it fairly closely, I allowed the damage to zombies to be doubled for double-damage bonus and boosted with spellcasting bonuses.

This should be pretty helpful with the rabid status. I also want to bring in the sick monsters flag and handling, so also quite nice to have with that threat.