Port diseased monsters behavior (from EvilHack).
This finishes up an important chapter for new monsters with disease attacks, namely locusts, gray fungi, and grave trolls. Before, their attacks would only affect the player, and the player wouldn't enjoy their special attack if polyd into their forms either.

Now, monsters can catch timed illness like the player and suffer a quick death shortly after contracting the disease. They also can use healing items to cure the illness.

This was also a good chance to catch a lot of stuff that I missed when creating the Rabid condition. Since illness and rabid are closely related, I found a lot of things to fix. Because we are not importing the sick/zombifying monsters from EvilHack, it was as simple as replacing msick with mrabid in most spots.

This also opens the door for a diseased weapons or artifacts. Grimtooth in EvilHack is pretty cool, but we won't copy it here and may try for something new.