Ported healthstones from SLASH'EM.
I think these will be a little more balanced in NerfHack than in SLASH'EM (or even HackEM).
- No tinkering recipes for these
- No guaranteed placements in the dungeon.

The 3.4.3 code for the healing calculations was very messy, but surprisingly in 3.7 they were super easy to bring in without much trouble. The 3.7 regeneration formula was tweaked and rebalanced recently, so I think this also helps these become a little less OP. If they don't seem to have much effect, we can add a small booster based on luck or something in the future.