New gray stone: foulstone.
I really enjoyed creating the stench property in HackEM - so this is a callback of sorts to that. A stinking item that nobody wants to be around!

Effects of carrying a foulstone:
* Causes aggravate monster (but without increasing the difficulty of spawned monsters)
* Monsters cannot digest you and you cannot digest monsters carrying one.
* Shopkeepers will not let you into their shops and won't buy it.
* If blessed, has a 1 in 100 chance each turn of emitting a stench that scares monsters for a short while (similar to garlic breath)
* If cursed, has a 1 in 100 chance each turn of emitting a poisonous cloud on your square.

Other effects:
* Monsters will never eat this rock and it's inedible if you are polyd into a monster that could eat it.
* Pets will try not to step on one (treated as a cursed item).
* If rubbed on another rock, emits a poisonous cloud.

In the previous commit 150193c7a, I changed Aggravate Monster to suppress monsters from fleeing. However, that conflicts with the mechanics for this item, so I partially reverted that commit so only musical instruments don't scare.