Put monster placement after dungeon feature gen (xNetHack).
Picked up some strange 'ceiling hider hiding in solid stone (fmon)' errors in the fuzzer. These were from piercers, so I suspect it has to do with the addition of trees in the main dungeon. I compared to xNetHack's mklev code and I missed putting monster placement after features, like trees. So I think it's a possible fix to piercers generating hidden on a tree, but not confirmed. In any case, moving that code is a good thing to do.