Monsters can zap wands of slow monster at the player.
This is adapted from EvilHack, but we don't have the uber-cool Slow status (yet). I'm not sure on that one because we are out of new status indicators on the bottom line. It's possible the Slow status indicator isn't mandatory for this and we can proceed with it anyway. This just gives monsters one more trick against the hero.