Allow monsters to wear rings.
Adapted from EvilHack. I just tried to do the bare minimum here and see what would work out of the box. Surprisingly, a lot of rings did function as expected.

- ring of invisibility (auto-ids if we see the effect)
- ring of fire resistance
- ring of cold resistance
- ring of poison resistance
- ring of shock resistance
- ring of protection

Rings with more complicated effects will need special handling in future commits.

Monsters will happily put on rings as long as you are a few spaces away and not a threat. There is not much for ring preference right now, but we have the foundation and can build from that.