Monsters can get extrinsic telepathy and shared telepathy.
The shared telepathy is a bit of an extra add-on from EvilHack, and it pops up in some other variants too. The list of telepathic monsters was greatly expanded using EvilHack's example as a guide. Because there are more source of telepathy - I have lowered the timeout to 150-400 turns to compensate. ESP should still be a difficult ability to acquire.

The amulet choice is getting a bit messy, but we will be fixing that in a future commit when we come to expanding extra_pref().