Flesh out monster preferences for item properties.
Finally we come to some monster AI! extra_pref is the function that values an item when a monster is considering it for wearing/wielding. Using EvilHack's established extra_pref as a template, I was able to adapt it without the object properties and streamline a few things here and there. Instead of manually checking each special wielded artifact and hero polyattack, I made a helper function to take care of that so we don't have to update this function everytime a new artifact is added or changed.

Stomping can be pretty dangerous, so I valued that fairly high for monsters.

I also touched up some missing things in muse.c for when monsters are searching for items.