Loyal steed attack fix (from EvilHack) and refactor pet targets.
This started out as an update patch from Evil, but turned into a refactor when I saw I had already cleaned up the pet targeting checked in HackEM.

From EvilHack commit f7c176351f0:
'Having a steed that was loyal or greater (tameness value of 15 or
greater) would let it attack on its own without waiting for you or it to
be attacked first. The problem was that if your steeds first target was
killed in that round, your steed could then go on to attack its next
target for free. And if it were also killed in that round, your steed
could then go on to its third target, and so on and so forth. This bug
has existed for over three years, but wasn't really noticed until
recently, when a player with a dragon steed watched it chain engulf
(which is an instant kill) four monsters in a row. I was able to
reproduce it and saw a chain of eight kills in a row in one round.

To fix it - redo the bit of code handling this behavior, and utilize
mspec_used to halt the steed from making another attack if its initial
target is killed. Your steed will still defend itself if it's attacked,
so you could still see 2-3 attacks from your steed in one round (it
attacks, kills its target, but in the same round it gets attacked by
other monsters and they land a hit, steed will then defend itself).

There may be a better way to resolve this, but this commit works, so
we'll roll with it for the time being.'