Ray-based wand of draining.
These are adapted from SLASH'EM, where they were beam based. I've been wanting to see a ray-based drain life wand since EvilHack introduced it with the shadow dragon (which is pretty cool!).

In SLASH'EM, this wand was notorious for being a bit overpowered and the bane of many adventurers. I'm hoping making it a ray wand makes it a little less threatening:
* It can be reflected, negating the major life draining.
* Monsters can also reflect it.
* When zapped over items on the floor, it can drain them like the drain life spell.
* Unfortunately, it can also bounce of walls - potentially making it more dangerous for the hero...
* Monsters will use it.