New monster: phoenix (from SpliceHack).
Also integrated many changes from Hack'EM to make these quite nasty to deal with.

When killed, phoenixes explode and leave behind a fresh phoenix egg.

This egg can hatch under many different circumstances:
* By hatching normally after a short period.
* By being exposed to fire or lava.
* If the player attempts to eat it one.
* If a pet attempts to eat one.
* If the player attempts to confine it in a container.
* If the player attempts to bury it (1 in 3 times the egg will hatch, but the other times it can be buried.)

The primary way to deal with phoenixes is to cancel them or their eggs. When cancelled, they will not be able to explode on death and will not leave an egg.

To make phoenixes not so annoying - I lowered their MC protection from 40 to 20. Their difficulty was also lowered from 20 to 15.

To enable the above effect, I also ported in: Zapping eggs with cancellation sterilizes them (from HackEM).