Fix polymorph behavior with extrinsic polymorphing monsters.
Monsters wearing a ring of polymorph will now act like SLASH'EM style shapeshifters. They will revert to their base form when killed.

I raised the odds of reversion for shapeshifters and werefoo from 1 in 7 to 1 in 13 to make it a bit more threatening. In SLASH'EM, the odds were 100% so this is still an improvement.

Getting this mechanic to work with the ring of protection from shapechangers was a little challenging but I believe we have it working now in parallel with the other shapeshifter behavior. Putting on a =oPfSS makes the monster go back to their base form while worn, and taking it off lets them poly again. When we kill the polyed monster, they always leave their base form's corpse.