Fix: water disintegrating too many scrolls.
This nerf from DynaHack was a bit to rough for my liking. I believe the intention was to destroy 1/3rd of scrolls dipped into a water source (or damaged by water), but for any stack it just destroyed the entire stack. This is not cool for blanking scrolls, so I adjusted it to take a less destructive random number when blanking stacks.