Clouds are all visible if outside the cloud.
I had reverted this before, but something must have changed upstream to undo it. Anyway, to reiterate my position on this: I much prefer gas clouds be visible than blocking vision. It takes away from the "video gameiness" if we just see nothing - especially if we just cast a scroll of stinking cloud...