Restrict card drops for revived/revivable monsters.
This is something I overlooked, but for zombies and trolls, a player could farm a lot of cards if they wanted. Same goes for the phoenix, which just keeps coming back over and over again. I also extended this to monsters revived via wand of undead turning and the like. This might also prevent a big of the card over-abundance in the early game, we'll see...