Implement hallucination items (from SLASH'EM).
This started as a fix for issue #1219 (https://github.com/NetHack/NetHack/issues/1219) which I submitted to the devteam. But I quickly realized there are many places in the code where items are simply not handled when hallucination. I thought back to the SLASH'EM way of handling this, and it's not as fancy as the bogus monsters are, but it is a good start and it really gives hallucination a unique spin on the game. You really have to remember some of your item keys to get a handle on it (or use a #pray).

The SLASH'EM way of handling randomized inventory is actually fairly clever. They simply intercept the item at xname, and then create a brand new item of the same class if you are hallucinating. This new item is simply swapped in and disposed of when the description is done. The downside is lack of variety, but I think that can be expanded on in the future.